Trying to add who won
Trying to add to see who won the match, but with csgo you can have overtime. So I will have to make a list and check through the list and see if any of the numbers match and if they do they win. So for example [16, 19, 22...] in a list and then is goes through and sees if it matches